refactor: CommentEditDialog  UI 분리
