+++ src/feature/comment/ui/index.ts
@@ -1,1 +1,2 @@
 export * from "./CommentAddDialog"
+export * from "./CommentEditDialog"
